Refactor the cache to live at the Source level
After moving the cache bulk of the cache lookup code into sdpcache
(see https://github.com/overmindtech/sdpcache/pull/64), this change now
shifts the cache into the individual sources. There is a new
`CachingSource` interface that provides access to the cache. The
SourceHost functions as dispatcher for Start/Stop/Purge at the engine
level.

Sources that want caching of their queries need to implement CachingSource
and provide a `sdpcache.Cache` instance on request.

All of this is to allow the source to access the cache locally when calling
into itself.