Use separate executions pols for LIST requests and everything else
Since LIST requests can block GET requests to load up the cache better,
there is a slight chance of a deadlock if too many get requests make it
into the pool before the LIST request. This creates a separate pool for
LIST requests to ensure that they can't deadlock waiting on the shared
pool.